+++ tracker/src/screens/screen_settings.c
@@ -1,4 +1,5 @@
 #include "screen_settings.h"
+#include "screen_color_theme.h"
 #include "common.h"
 #include "corelib_gfx.h"
 #include "corelib_mainloop.h"
@@ -14,7 +15,7 @@ static void settingsDrawField(int col, int row, int state);
 static int settingsOnEdit(int col, int row, enum CellEditAction action);
 
 static ScreenData screenSettingsData = {
-  .rows = 5,
+  .rows = 6,
   .cursorRow = 0,
   .cursorCol = 0,
   .selectMode = -1,
@@ -58,6 +59,8 @@ void settingsDrawCursor(int col, int row) {
   } else if (row == 3 && col == 0) {
     gfxCursor(23, 5, 3); // Under gamepad swap ON/OFF
   } else if (row == 4 && col == 0) {
+    gfxCursor(0, 6, 16); // "Edit color theme"
+  } else if (row == 5 && col == 0) {
     gfxCursor(0, 17, 14); // "Quit ChipNomad"
   }
 }
@@ -93,6 +96,9 @@ void settingsDrawField(int col, int row, int state) {
     gfxPrint(23, 5, appSettings.gamepadSwapAB ? "ON " : "OFF");
   } else if (row == 4 && col == 0) {
     gfxSetFgColor(state == stateFocus ? appSettings.colorScheme.textValue : appSettings.colorScheme.textDefault);
+    gfxPrint(0, 6, "Edit color theme");
+  } else if (row == 5 && col == 0) {
+    gfxSetFgColor(state == stateFocus ? appSettings.colorScheme.textValue : appSettings.colorScheme.textDefault);
     gfxPrint(0, 17, "Quit ChipNomad");
   }
 }
@@ -127,6 +133,10 @@ int settingsOnEdit(int col, int row, enum CellEditAction action) {
     static uint8_t lastValue = 0;
     return edit8withLimit(action, (uint8_t*)&appSettings.gamepadSwapAB, &lastValue, 1, 1);
   } else if (row == 4 && col == 0 && action == editTap) {
+    // Navigate to color theme screen
+    screenSetup(&screenColorTheme, 0);
+    return 0;
+  } else if (row == 5 && col == 0 && action == editTap) {
     // Trigger exit event
     mainLoopTriggerQuit();
     return 1;