+++ tracker/src/screens/screens.c
@@ -158,6 +158,7 @@ static void validateCursorBounds(ScreenData* screen) {
 
 void screenFullRedraw(ScreenData* screen) {
   validateCursorBounds(screen);
+  gfxSetBgColor(appSettings.colorScheme.background);
   gfxClearRect(0, 0, 40, 20);
 
   // Static content
@@ -479,7 +480,7 @@ static int inputSelectMode(ScreenData* screen, int keys, int isDoubleTap) {
 int screenInput(ScreenData* screen, int isKeyDown, int keys, int isDoubleTap) {
   // Discard key up events unless no buttons are pressed (for existing logic that expects keys == 0)
   if (!isKeyDown && keys != 0) return 0;
-  
+
   return (screen->selectMode == 1) ? inputSelectMode(screen, keys, isDoubleTap) : inputNormalMode(screen, keys, isDoubleTap);
 }