+++ tracker/src/screens/screens.h
@@ -76,6 +76,7 @@ extern const AppScreen screenTable;
 extern const AppScreen screenExport;
 extern const AppScreen screenManage;
 extern const AppScreen screenSettings;
+extern const AppScreen screenColorTheme;
 
 extern const AppScreen* currentScreen;