+++ CHANGELOG.md
@@ -3,17 +3,23 @@
 ## Not yet released
 
 - *PLATFORM*: Linux x86_64 package for Linux desktops and Steam Deck
+- Gamepad support for desktop builds
 - Support for 2x and 3x AY/YM chips
 - Linear pitch option (pitch tables are defined in cents)
-- Mute/solo tracks
+- Mute/solo tracks (B + Select/Start on Song screen. Release B first to keep mute/solo)
 - Clean up of unused instruments, unused/duplicate phrases and chains
+- Chain and Phrase screens show asterisk next to chain/phrase number if it's used elsewhere in the song
 - B+A on an empty cell at the Song screen now moves the whole column up (same as in LSDJ and M8)
 - Project and Instrument save functions now check for empty filename before saving
 - AY/YM emulator filter quality setting (lower quality - lower CPU load)
 - Looping cursor in the file browser
+- Color theme edit, load, and save
 - *FIX*: Chip settings were not initialized when loading a project
 - *FIX*: UI was monochrome in RG35xx build
-- *FIX*: All saved values are correctly reset on loading or creating a new project
+- *FIX*: All saved values are correctly reset on loading or creating a new project now
+- *FIX*: Multi edit bug on FX columns in phrases and tables
+- *FIX*: Project title is lost when you load a project with an empty author
+- *FIX*: Chain deep clone created one more copy of a chain
 
 ## v0.0.3a (November 22, 2025)